03-Apps/07-Legal_Document_Analyzer_Agent/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.40
openai>=1.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
pillow>=9.0.0
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
typing-extensions>=4.5.0